Bug 1197186 - Move whitenoise_custom.py to treeherder/config/
Since it's not specific to the Django app 'webapp'.
