+++ treeherder/config/wsgi.py
@@ -17,7 +17,7 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.settings")
 from django.core.cache.backends.memcached import BaseMemcachedCache
 from django.core.wsgi import get_wsgi_application
 
-from treeherder.webapp.whitenoise_custom import CustomWhiteNoise
+from treeherder.config.whitenoise_custom import CustomWhiteNoise
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
